# docs/swagger.yaml
definitions:
  handlers.CreateDocumentRequest:
    properties:
      description:
        type: string
      is_public:
        type: boolean
      title:
        type: string
    type: object
  handlers.DocumentListResponse:
    properties:
      owned:
        items:
          $ref: '#/definitions/handlers.DocumentResponse'
        type: array
      shared:
        items:
          $ref: '#/definitions/handlers.DocumentResponse'
        type: array
    type: object
  handlers.DocumentResponse:
    properties:
      content:
        items:
          type: integer
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      owner_id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  handlers.DocumentVersionResponse:
    properties:
      changed_by:
        type: string
      content:
        items:
          type: integer
        type: array
      created_at:
        type: string
      document_id:
        type: string
      id:
        type: string
      version:
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.UpdateDocumentRequest:
    properties:
      content:
        items:
          type: integer
        type: array
      description:
        type: string
      is_public:
        type: boolean
      status:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/documents:
    post:
      consumes:
      - application/json
      description: Create a new document with title, description, and content
      parameters:
      - description: Document Data
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Document created successfully
          schema:
            $ref: '#/definitions/handlers.DocumentResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new document
      tags:
      - Documents
  /api/v1/documents/{id}:
    get:
      description: Retrieve a document by its unique identifier
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document retrieved successfully
          schema:
            $ref: '#/definitions/handlers.DocumentResponse'
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a document by ID
      tags:
      - Documents
    put:
      consumes:
      - application/json
      description: Update the title, description, or content of an existing document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated document data
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Document updated successfully
          schema:
            $ref: '#/definitions/handlers.DocumentResponse'
        "400":
          description: Invalid document ID or request format
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update an existing document
      tags:
      - Documents
  /api/v1/documents/{id}/versions:
    get:
      description: Retrieve all versions of a specific document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Document version history
          schema:
            items:
              $ref: '#/definitions/handlers.DocumentVersionResponse'
            type: array
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get version history of a document
      tags:
      - Documents
  /api/v1/documents/user:
    get:
      description: Retrieve all owned and shared documents for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Documents retrieved successfully
          schema:
            $ref: '#/definitions/handlers.DocumentListResponse'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all documents for the authenticated user
      tags:
      - Documents
...
swagger: "2.0"